+++ doc/changelog.rst
@@ -24,6 +24,13 @@ API Changes
   ``pcca.microstate_mapping_`` for this functionality. 
 - ``UnionDataset`` has been removed after deprecation in 3.3. Please use
   ``FeatureUnion`` instead (#671).
+- ``SubsetFeaturizer`` and ilk have been removed from the
+  ``msmbuilder.featurizer`` namespace. Please import them from
+  ``msmbuilder.featurizer.subset`` (#738).
+- ``FirstSlicer`` has been removed. Use ``Slicer(first=x)`` for the same
+  functionality (#738).
+- ``msmbuilder.featurizer.load`` has been removed. ``Featurizer.save``
+  has been removed. Please use ``utils.load``, ``utils.dump`` (#738).
 
 
 New Features